Address Carter's comments
