+++ doc/usage/al2023.md
@@ -46,3 +46,22 @@
 | `volume_type` |  |
 | `working_dir` | Directory path for ephemeral resources on the builder instance |
 <!-- template-variable-table-boundary -->
+
+## Accelerated images
+
+One can build images that contain Neuron and Nvidia drivers and runtime configuration. To build Neuron image execute:
+
+```
+make k8s=1.29 os_distro=al2023 accelerator_vendor=neuron
+```
+
+To build Nvidia image execute:
+```
+make k8s=1.29 os_distro=al2023 accelerator_vendor=nvidia
+```
+
+One can pass the Nvidia major driver version using the following:
+```
+make k8s=1.29 os_distro=al2023 accelerator_vendor=nvidia nvidia_major_driver_version=560
+```
+To see which driver versions are available, one can check the Nvidia AL2023 yum [repository](https://developer.download.nvidia.com/compute/cuda/repos/amzn2023/).
