+++ nodeadm/test/e2e/cases/containerd-runtime-config-nvidia/expected-containerd-config.toml
@@ -19,13 +19,12 @@ discard_unpacked_layers = false
 
 [plugins.'io.containerd.grpc.v1.cri'.containerd.runtimes]
 [plugins.'io.containerd.grpc.v1.cri'.containerd.runtimes.nvidia]
-privileged_without_host_devices = false
-runtime_engine = ''
-runtime_root = ''
+base_runtime_spec = '/etc/containerd/base-runtime-spec.json'
 runtime_type = 'io.containerd.runc.v2'
 
 [plugins.'io.containerd.grpc.v1.cri'.containerd.runtimes.nvidia.options]
 BinaryName = '/usr/bin/nvidia-container-runtime'
+SystemdCgroup = true
 
 [plugins.'io.containerd.grpc.v1.cri'.containerd.runtimes.runc]
 base_runtime_spec = '/etc/containerd/base-runtime-spec.json'
