--- templates/al2023/runtime/gpu/bootstrap_gpu.service
@@ -1,11 +1,0 @@
-[Unit]
-Description=Nvidia Bootstrap Service
-After=network.target nodeadm-config.service
-
-[Service]
-Type=oneshot
-ExecStart=/etc/eks/bootstrap_gpu.sh
-RemainAfterExit=true
-
-[Install]
-WantedBy=multi-user.target
