+++ templates/al2023/runtime/gpu/initialize-nvidia-clock.sh
