+++ templates/al2023/runtime/gpu/nvidia-kmod-load.sh
