+++ templates/al2023/template.json
@@ -236,6 +236,15 @@
     {
       "type": "shell",
       "remote_folder": "{{ user `remote_folder`}}",
+      "script": "{{template_dir}}/provisioners/configure-nvidia-eula.sh",
+      "environment_vars": [
+        "ACCELERATOR_VENDOR={{user `accelerator_vendor`}}",
+        "WORKING_DIR={{user `working_dir`}}"
+      ]
+    },
+    {
+      "type": "shell",
+      "remote_folder": "{{ user `remote_folder`}}",
       "script": "{{template_dir}}/provisioners/install-efa.sh",
       "environment_vars": [
         "AWS_REGION={{user `aws_region`}}",
