+++ templates/al2023/variables-default.json
@@ -18,7 +18,7 @@
     "kms_key_id": "",
     "iam_instance_profile": "",
     "launch_block_device_mappings_volume_size": "20",
-    "nvidia_major_driver_version": "",
+    "nvidia_major_driver_version": "555",
     "remote_folder": "/tmp",
     "runc_version": "*",
     "security_group_id": "",
